Move angular services into their own folderb
