+++ app/assets/js/services/index.js
@@ -1,0 +1,3 @@
+angular.module('nitServices', ['LocalStorageModule'])
+  .factory('articleService', require('./article-service'))
+  .factory('unionService', require('./union-service'));
